refactor(service): remove msgpack from standard package
