+++ docs/en_US/operation/compile/features.md
@@ -12,6 +12,7 @@ Except core runtime and REST api, there are some features that are allowed to be
 | [Native plugin](../../extension/native/overview.md)                                               | plugin     | The native plugin runtime, REST API, CLI API etc.                                                                                                      |
 | [Portable plugin](../../extension/portable/overview.md)                                           | portable   | The portable plugin runtime, REST API, CLI API etc.                                                                                                    |
 | [External service](../../extension/external/external_func.md)                                     | service    | The external service runtime, REST API, CLI API etc.                                                                                                   |
+| [Msgpack-rpc External service](../../extension/external/external_func.md)                         | msgpack    | Support msgpack-rpc protocol in external service                                                                                                       |
 | [UI Meta API](../../operation/manager-ui/overview.md)                                             | ui         | The REST API of the metadata which is usually consumed by the ui                                                                                       |
 | [Prometheus Metrics](../../configuration/global_configurations.md#prometheus-configuration)       | prometheus | Support to send metrics to prometheus                                                                                                                  |
 | [Extended template functions](../../guide/sinks/data_template.md#functions-supported-in-template) | template   | Support additional data template function from sprig besides default go text/template functions                                                        |
