+++ internal/service/executors.go
@@ -1,4 +1,4 @@
-// Copyright 2021-2022 EMQ Technologies Co., Ltd.
+// Copyright 2021-2023 EMQ Technologies Co., Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,20 +19,15 @@ import (
 	"crypto/tls"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
-	"net/rpc"
 	"net/url"
-	"reflect"
 	"strings"
-	"sync"
 	"time"
 
 	// TODO: replace with `google.golang.org/protobuf/proto` pkg.
 	"github.com/golang/protobuf/proto" //nolint:staticcheck
 	"github.com/jhump/protoreflect/dynamic"
 	"github.com/jhump/protoreflect/dynamic/grpcdynamic"
-	"github.com/ugorji/go/codec"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 
@@ -42,8 +37,45 @@ import (
 	"github.com/lf-edge/ekuiper/pkg/infra"
 )
 
+type exeIns func(desc descriptor, opt *interfaceOpt, i *interfaceInfo) (executor, error)
+
+var executors = map[protocol]exeIns{
+	GRPC: newGrpcExecutor,
+	REST: newHttpExecutor,
+}
+
+func newHttpExecutor(desc descriptor, opt *interfaceOpt, i *interfaceInfo) (executor, error) {
+	d, ok := desc.(multiplexDescriptor)
+	if !ok {
+		return nil, fmt.Errorf("invalid descriptor type for rest")
+	}
+	o := &restOption{}
+	e := cast.MapToStruct(i.Options, o)
+	if e != nil {
+		return nil, fmt.Errorf("incorrect rest option: %v", e)
+	}
+	exe := &httpExecutor{
+		descriptor:   d,
+		interfaceOpt: opt,
+		restOpt:      o,
+	}
+	return exe, nil
+}
+
+func newGrpcExecutor(desc descriptor, opt *interfaceOpt, _ *interfaceInfo) (executor, error) {
+	d, ok := desc.(protoDescriptor)
+	if !ok {
+		return nil, fmt.Errorf("invalid descriptor type for grpc")
+	}
+	exe := &grpcExecutor{
+		descriptor:   d,
+		interfaceOpt: opt,
+	}
+	return exe, nil
+}
+
 // NewExecutor
-// Each interface definition maps to one executor instance. It is suppose to have only one thread running.
+// Each interface definition maps to one executor instance. It is supposed to have only one thread running.
 func NewExecutor(i *interfaceInfo) (executor, error) {
 	// No validation here, suppose the validation has been done in json parsing
 	descriptor, err := parse(i.Schema.SchemaType, i.Schema.SchemaFile)
@@ -58,44 +90,10 @@ func NewExecutor(i *interfaceInfo) (executor, error) {
 		addr:    u,
 		timeout: 5000,
 	}
-	switch i.Protocol {
-	case GRPC:
-		d, ok := descriptor.(protoDescriptor)
-		if !ok {
-			return nil, fmt.Errorf("invalid descriptor type for grpc")
-		}
-		exe := &grpcExecutor{
-			descriptor:   d,
-			interfaceOpt: opt,
-		}
-		return exe, nil
-	case REST:
-		d, ok := descriptor.(multiplexDescriptor)
-		if !ok {
-			return nil, fmt.Errorf("invalid descriptor type for rest")
-		}
-		o := &restOption{}
-		e := cast.MapToStruct(i.Options, o)
-		if e != nil {
-			return nil, fmt.Errorf("incorrect rest option: %v", e)
-		}
-		exe := &httpExecutor{
-			descriptor:   d,
-			interfaceOpt: opt,
-			restOpt:      o,
-		}
-		return exe, nil
-	case MSGPACK:
-		d, ok := descriptor.(interfaceDescriptor)
-		if !ok {
-			return nil, fmt.Errorf("invalid descriptor type for msgpack-rpc")
-		}
-		exe := &msgpackExecutor{
-			descriptor:   d,
-			interfaceOpt: opt,
-		}
-		return exe, nil
-	default:
+
+	if ins, ok := executors[i.Protocol]; ok {
+		return ins(descriptor, opt, i)
+	} else {
 		return nil, fmt.Errorf("unsupported protocol %s", i.Protocol)
 	}
 }
@@ -237,58 +235,4 @@ func (h *httpExecutor) InvokeFunction(ctx api.FunctionContext, name string, para
 			return nil, fmt.Errorf("unsupported resposne content type %s", contentType)
 		}
 	}
-}
-
-type msgpackExecutor struct {
-	descriptor interfaceDescriptor
-	*interfaceOpt
-
-	sync.Mutex
-	connected bool
-	conn      *rpc.Client
-}
-
-// InvokeFunction flat the params and result
-func (m *msgpackExecutor) InvokeFunction(_ api.FunctionContext, name string, params []interface{}) (interface{}, error) {
-	if !m.connected {
-		m.Lock()
-		if !m.connected {
-			h := &codec.MsgpackHandle{}
-			h.MapType = reflect.TypeOf(map[string]interface{}(nil))
-
-			conn, err := net.Dial(m.addr.Scheme, m.addr.Host)
-			if err != nil {
-				return nil, err
-			}
-			rpcCodec := codec.MsgpackSpecRpc.ClientCodec(conn, h)
-			m.conn = rpc.NewClientWithCodec(rpcCodec)
-		}
-		m.connected = true
-		m.Unlock()
-	}
-	ps, err := m.descriptor.ConvertParams(name, params)
-	if err != nil {
-		return nil, err
-	}
-	var (
-		reply interface{}
-		args  interface{}
-	)
-	// TODO argument flat
-	switch len(ps) {
-	case 0:
-		// do nothing
-	case 1:
-		args = ps[0]
-	default:
-		args = codec.MsgpackSpecRpcMultiArgs(ps)
-	}
-	err = m.conn.Call(name, args, &reply)
-	if err != nil {
-		if err == rpc.ErrShutdown {
-			m.connected = false
-		}
-		return nil, err
-	}
-	return m.descriptor.ConvertReturn(name, reply)
 }
