+++ internal/service/external_service_rule_test.go
@@ -29,7 +29,6 @@ import (
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/golang/protobuf/ptypes/wrappers"
 	"github.com/gorilla/mux"
-	"github.com/msgpack-rpc/msgpack-rpc-go/rpc"
 	"google.golang.org/grpc"
 
 	kconf "github.com/lf-edge/ekuiper/internal/conf"
@@ -579,135 +578,6 @@ func object_detection(command string, image string) map[string]interface{} {
 
 func getStatus() bool {
 	return true
-}
-
-func TestMsgpackService(t *testing.T) {
-	// mock server
-	res := Resolver{"SayHello": reflect.ValueOf(SayHello), "object_detection": reflect.ValueOf(object_detection), "get_feature": reflect.ValueOf(get_feature), "getStatus": reflect.ValueOf(getStatus)}
-	serv := rpc.NewServer(res, true, nil)
-	l, _ := net.Listen("tcp", ":50000")
-	serv.Listen(l)
-	go serv.Run()
-	// Comment out because the bug in the msgpack rpc
-	// defer serv.Stop()
-
-	// Reset
-	streamList := []string{"helloStr", "commands", "fakeBin"}
-	topotest.HandleStream(false, streamList, t)
-	// Data setup
-	tests := []topotest.RuleTest{
-		{
-			Name: `TestRestRule1`,
-			Sql:  `SELECT helloFromMsgpack(name) as wc FROM helloStr`,
-			R: [][]map[string]interface{}{
-				{{
-					"wc": map[string]interface{}{
-						"message": "world",
-					},
-				}},
-				{{
-					"wc": map[string]interface{}{
-						"message": "golang",
-					},
-				}},
-				{{
-					"wc": map[string]interface{}{
-						"message": "peacock",
-					},
-				}},
-			},
-			M: map[string]interface{}{
-				"op_2_project_0_exceptions_total":   int64(0),
-				"op_2_project_0_process_latency_us": int64(0),
-				"op_2_project_0_records_in_total":   int64(3),
-				"op_2_project_0_records_out_total":  int64(3),
-
-				"sink_mockSink_0_exceptions_total":  int64(0),
-				"sink_mockSink_0_records_in_total":  int64(3),
-				"sink_mockSink_0_records_out_total": int64(3),
-			},
-		}, {
-			Name: `TestRestRule2`,
-			Sql:  `SELECT objectDetectFromMsgpack(*)->result FROM commands`,
-			R: [][]map[string]interface{}{
-				{{
-					"kuiper_field_0": "get success",
-				}},
-				{{
-					"kuiper_field_0": "detect success",
-				}},
-				{{
-					"kuiper_field_0": "delete success",
-				}},
-			},
-			M: map[string]interface{}{
-				"op_2_project_0_exceptions_total":   int64(0),
-				"op_2_project_0_process_latency_us": int64(0),
-				"op_2_project_0_records_in_total":   int64(3),
-				"op_2_project_0_records_out_total":  int64(3),
-
-				"sink_mockSink_0_exceptions_total":  int64(0),
-				"sink_mockSink_0_records_in_total":  int64(3),
-				"sink_mockSink_0_records_out_total": int64(3),
-			},
-		}, {
-			Name: `TestRestRule3`,
-			Sql:  `SELECT getFeatureFromMsgpack(self)->feature[0]->box->h FROM fakeBin`,
-			R: [][]map[string]interface{}{
-				{{
-					"kuiper_field_0": float64(106), // Convert by the testing tool
-				}},
-				{{
-					"kuiper_field_0": float64(107),
-				}},
-				{{
-					"kuiper_field_0": float64(108),
-				}},
-			},
-			M: map[string]interface{}{
-				"op_2_project_0_exceptions_total":   int64(0),
-				"op_2_project_0_process_latency_us": int64(0),
-				"op_2_project_0_records_in_total":   int64(3),
-				"op_2_project_0_records_out_total":  int64(3),
-
-				"sink_mockSink_0_exceptions_total":  int64(0),
-				"sink_mockSink_0_records_in_total":  int64(3),
-				"sink_mockSink_0_records_out_total": int64(3),
-			},
-			//}, {
-			//	Name: `TestRestRule4`,
-			//	Sql:  `SELECT getStatusFromMsgpack(), command FROM commands`,
-			//	R: [][]map[string]interface{}{
-			//		{{
-			//			"getStatusFromRest": true,
-			//			"command": "get",
-			//		}},
-			//		{{
-			//			"getStatusFromRest": true,
-			//			"command": "detect",
-			//		}},
-			//		{{
-			//			"getStatusFromRest": true,
-			//			"command": "delete",
-			//		}},
-			//	},
-			//	M: map[string]interface{}{
-			//		"op_2_project_0_exceptions_total":   int64(0),
-			//		"op_2_project_0_process_latency_us": int64(0),
-			//		"op_2_project_0_records_in_total":   int64(3),
-			//		"op_2_project_0_records_out_total":  int64(3),
-			//
-			//		"sink_mockSink_0_exceptions_total":  int64(0),
-			//		"sink_mockSink_0_records_in_total":  int64(3),
-			//		"sink_mockSink_0_records_out_total": int64(3),
-			//	},
-		},
-	}
-	topotest.HandleStream(true, streamList, t)
-	topotest.DoRuleTest(t, tests, 0, &api.RuleOption{
-		BufferLength: 100,
-		SendError:    true,
-	}, 0)
 }
 
 type server struct {
